Fix files larger than 2.15 GB can't be saved (fix #5935)
Before this fix the max file size without issues was 2.15GB,
now, Aseprite files can now exceed the current practical capacity
of hard drives.

Note: The on-disk header size field remains a uint32_t,
so it overflows for files > 4.30 GB. The only practical consequence
is that the progress bar becomes inaccurate when opening
files > 4.30 GB.

Also limit palette creation to 16384 colors when an RGBA sprite
contains a Black palette, preventing Aseprite from hanging due
to millions of unique colors in very large sprites.

+++ src/desktop/win/thumbnail_handler.cpp
@@ -50,7 +50,7 @@ public:
 
   bool ok() const override { return m_ok; }
 
-  size_t tell() override
+  uint64_t tell() override
   {
     LARGE_INTEGER delta;
     delta.QuadPart = 0;
@@ -64,7 +64,7 @@ public:
     return newPos.QuadPart;
   }
 
-  void seek(size_t absPos) override
+  void seek(uint64_t absPos) override
   {
     LARGE_INTEGER pos;
     pos.QuadPart = absPos;
